post vehicle api created

+++ vehicle-management-service/src/main/resources/db/migration/V5__create_customer_vehicles_table.sql
@@ -10,7 +10,7 @@ CREATE TABLE customer_vehicles
 
     registration_number VARCHAR(20) NOT NULL,
 
-    manufacturing_year SMALLINT,
+    manufacturing_year INTEGER,
 
     color VARCHAR(50),
